requirements.txt
langchain
langgraph
langchain-community
langchain-google-genai
langchain-groq
chromadb
python-multipart
sentence-transformers
pypdf
fastapi
uvicorn
streamlit
weasyprint
jinja2
duckduckgo-search
tavily-python
python-dotenv
pydantic